+++ src/test-komodo/testutils.h
@@ -34,4 +34,12 @@ void deleteIfUsedBefore(T*& ptr) {
     }
 }
 
+// C++ stream interface
+class TestCout : public std::stringstream {
+    public:
+        ~TestCout();
+};
+
+#define GTEST_COUT_COLOR TestCout()
+#define GTEST_COUT_NOCOLOR std::cerr << "[          ] [ INFO ] "
 #endif /* TESTUTILS_H */
